requirements.txt
# VisionOps 服务器端依赖

# ── 核心训练框架 ──
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.0        # YOLOv8

# ── MLOps ──
mlflow>=2.13.0
dvc[s3]>=3.0.0            # DVC + S3(MinIO)存储后端
hydra-core>=1.3.0         # 配置管理

# ── ONNX导出与优化 ──
onnx>=1.15.0
onnxruntime>=1.17.0
onnxsim>=0.4.0            # onnx-simplifier

# ── Web后端 ──
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── 数据处理 ──
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0

# ── 数据库 ──
psycopg2-binary>=2.9.0
boto3>=1.34.0             # S3/MinIO客户端

# ── 工具 ──
pyyaml>=6.0.0
tqdm>=4.66.0
requests>=2.31.0
python-dotenv>=1.0.0
